requirements.txt
colorlog==6.8.0
pyyaml==6.0.1
beautifulsoup4==4.12.2
